Address review feedback
This changes the actions taken when a cache cannot be found to
waiting up to 90s for the file to be in the cache before redirecting
back to the original url.  This also splits the backends into two
different files so that it's easier to see what would need to be
done to implement a new backend class

It also updates the Dockerfile to node 5.  Because of how we're
deploying (tutum instead of heroku) we need this dockerfile
for both sides of the deployment